Removed SDL.h and windows.h global includes
