+++ src/audio/sampled_stream.c
@@ -22,6 +22,8 @@
 
 // Common functions for sampled streams.
 
+#include <math.h>
+
 #include "audio.h"
 #include "sampled_stream.h"
